Update plugin2.ts
We parse the file that is going to be passed to `bun run` in the preloaded script, before we call `plugin()`. That is, we do what we want in the preload script outside of just relying on `bun.plugin()` to do the work. We use `RegExp` to match specifiers begining an arbitrary `specifier:` specifier, replace the exported content.